Pull latest changes from ref using cached repo
Act switched from checking out a branch to checking out
a hash in order to resolve #24 - see
bc5c23e8e48ee1886ac274965852196d1ec3ba3f

This had a side effect of preventing `w.Pull()` from updating
to the latest commit if a ref was a branch/tag and the repo
already exists on disk.

This commit checks if input.Ref is equivalent to the resolved
hash, and runs `r.Fetch()` and checks out the branch for input.Ref
if it is different. `w.Pull()` can then pull the latest changes
before we checkout `*hash` again like normal