feat: try GITHUB_TOKEN > SSH agent > unauthenticated git clone
This makes it easier to run act locally with GitHub enterprise, because
cloning remote actions otherwise always requires setting `GITHUB_TOKEN`.

It will also help with remote actions in private repositories.

Unfortunately a `GITHUB_TOKEN` still needs to be provided for the
`actions/checkout` action when cloning a remote repository or different
ref (i.e. not a local checkout). This *might* change, if:
https://github.com/actions/checkout/discussions/660 is acceptable.

Also a token needs to be provided in case the workflow contains actions
that talk to the GitHub API.