refactor: remove `NewExpressionEvaluator()` from `StepContext`
Except for the `exprparser.Config.Context` (which can be either "job"
or "step"), both expression evaluators (on the `RunContext` and on the
`StepContext`) were identical.
In the future we might want to have different evaluators, because the
`Steps` and `Inputs` should be different based on the context, but at
the moment both evaluators need access to all inputs in order to allow
composite actions to function properly.

This commit removes the `NewExpressionEvaluator()` method on the
`StepContext` and uses `NewExpressionEvaluator("step")` when a step
evaluator is needed.