Update README with note about docker login ghcr.io
As discussed in https://github.com/nektos/act/issues/1248 , the images for this tool are now hosted in `ghcr.io` which (apparently) no longer allows unauthenticated pulls. This PR adds some documentation around how to login to `ghcr.io` as well as mentions the error that occurs if you are not in order to help new users get over this stumbling block.

New contributor here, just wanted to say thanks to all the contributors for making this project. ❤️ I'm looking forward to a large reduction in the number of `fix: fmt` and `fix: linting` commits in my PRs now that I can run my workflows locally. 😂